changed potentiometer int to float

+++ 2015/src/org/usfirst/frc/team2972/robot/subsystems/Turret.java
@@ -39,7 +39,7 @@ public class Turret extends Subsystem {
     	turretMotor.set(0);
     }
     
-    public int getPotentiometerValue(){
+    public float getPotentiometerValue(){
     	return potentiometer.getValue();
     }
